@@ -1,7 +1,9 @@
-package com.example.demowithtests.util;
+package com.example.demowithtests.util.exception;
+
+import lombok.Getter;
 
 import java.util.Date;
-
+@Getter
 public class ErrorDetails {
     private Date timestamp;
     private String message;
@@ -12,17 +14,5 @@ public class ErrorDetails {
         this.timestamp = timestamp;
         this.message = message;
         this.details = details;
-    }
-
-    public Date getTimestamp() {
-        return timestamp;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public String getDetails() {
-        return details;
     }
 }
